# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SafePay API
  version: "1.0.0"
  description: |
    **Trusted payments, everywhere.**

    SafePay is an escrow, settlement and trust-scoring layer for Nigeria. Any
    application — a marketplace, a freelance platform, a storefront, a landlord's
    portal — can plug in so that a transaction between two people who do not fully
    trust each other becomes safe by default.

    ## Authentication

    Two credentials reach the same endpoints:

    | Credential | Header | Used by |
    |---|---|---|
    | Dashboard session (JWT) | `Authorization: Bearer eyJ…` | The SafePay web app |
    | API key | `Authorization: Bearer sk_test_…` | Partner applications |

    Keys come in pairs. `sk_test_` keys operate on simulated settlement — build the
    entire flow, including disputes and webhooks, without moving a naira.
    SafePay stores only a SHA-256 hash of every key, so a lost key is rotated,
    never recovered.

    ## Money

    Every amount crosses the wire as an **integer in kobo** (`amountKobo`). Endpoints
    that create an escrow also accept `amount` in naira as a convenience, and convert
    it. Floating point never touches a balance.

    ## Errors

    Failures return a consistent envelope:

    ```json
    { "error": { "code": "conflict", "message": "An escrow that is \"released\" cannot move to \"released\"." } }
    ```

    ## Rate limits

    Responses carry `RateLimit-Limit`, `RateLimit-Remaining` and `RateLimit-Reset`.
...
  contact:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /v1/score/{userId}:
    get:
      tags: [SafeScore]
      summary: Look up a SafeScore
      description: |
        **Public — no authentication required.** This is reputation portability: any
        third-party application can check a seller before money changes hands, the way
        a credit bureau works, but for everyday trading.

        Accepts a user id or an email address. The full component breakdown and total
        value settled are returned only to the score's owner or an admin.
      security: []
      parameters:
        - { name: userId, in: path, required: true, schema: { type: string }, example: usr_JbRmBoCmk1JrSKft }
      responses:
        "200":
          description: The public trust profile.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicScore" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/score/{userId}/badge.svg:
    get:
      tags: [SafeScore]
      summary: Embeddable Trust Badge
      description: |
        Returns a self-contained SVG. Drop it anywhere an image works — an Instagram
        bio link, a WhatsApp storefront, a website — with no SDK and no JavaScript:

        ```html
        <img src="https://api.safepay.ng/v1/score/usr_123/badge.svg" width="300" height="76">
        ```
      security: []
      parameters:
        - { name: userId, in: path, required: true, schema: { type: string } }
        - { name: theme, in: query, schema: { type: string, enum: [light, dark], default: light } }
      responses:
        "200":
          description: An SVG badge, cacheable for 5 minutes.
          content:
            image/svg+xml:
              schema: { type: string }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: A dashboard session JWT, or an API key beginning `sk_test_` / `sk_live_`.
...
  responses:
...
    NotFound:
      description: No such resource.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: conflict }
            message: { type: string }
            details: { type: object }
...
    PublicScore:
      type: object
      properties:
        user:
          type: object
          properties:
            id: { type: string }
            name: { type: string }
            memberSince: { type: string, format: date-time }
        score: { type: integer, example: 68 }
        tier: { type: string }
        tierLabel: { type: string }
        stats:
          type: object
          properties:
            escrowsCompleted: { type: integer }
            disputeRatePct: { type: number }
            uniqueCounterparties: { type: integer }
            verificationTier: { type: string }
            medianReleaseHours: { type: number, nullable: true }
        updatedAt: { type: string, format: date-time }
...